OP-2042 - added additional class wrapper per host
