+++ tests/integration/hosts/maya/test_publish_in_maya.py
@@ -1,11 +1,7 @@
-import pytest
-import os
-import shutil
-
-from tests.lib.testing_classes import PublishTest
+from tests.integration.hosts.maya.lib import MayaTestClass
 
 
-class TestPublishInMaya(PublishTest):
+class TestPublishInMaya(MayaTestClass):
     """Basic test case for publishing in Maya
 
         Shouldnt be running standalone only via 'runtests' pype command! (??)
@@ -37,40 +33,6 @@ class TestPublishInMaya(PublishTest):
     APP_VARIANT = ""
 
     TIMEOUT = 120  # publish timeout
-
-    @pytest.fixture(scope="module")
-    def last_workfile_path(self, download_test_data, output_folder_url):
-        """Get last_workfile_path from source data.
-
-            Maya expects workfile in proper folder, so copy is done first.
-        """
-        src_path = os.path.join(download_test_data,
-                                "input",
-                                "workfile",
-                                "test_project_test_asset_TestTask_v001.mb")
-        dest_folder = os.path.join(output_folder_url,
-                                   self.PROJECT,
-                                   self.ASSET,
-                                   "work",
-                                   self.TASK)
-        os.makedirs(dest_folder)
-        dest_path = os.path.join(dest_folder,
-                                 "test_project_test_asset_TestTask_v001.mb")
-        shutil.copy(src_path, dest_path)
-
-        yield dest_path
-
-    @pytest.fixture(scope="module")
-    def startup_scripts(self, monkeypatch_session, download_test_data):
-        """Points Maya to userSetup file from input data"""
-        startup_path = os.path.join(download_test_data,
-                                    "input",
-                                    "startup")
-        original_pythonpath = os.environ.get("PYTHONPATH")
-        monkeypatch_session.setenv("PYTHONPATH",
-                                   "{}{}{}".format(startup_path,
-                                                   os.pathsep,
-                                                   original_pythonpath))
 
     def test_db_asserts(self, dbcon, publish_finished):
         """Host and input data dependent expected results in DB."""
