+++ tests/integration/hosts/nuke/test_publish_in_nuke.py
@@ -1,15 +1,12 @@
-import pytest
-import os
 import logging
-import shutil
 
-from tests.lib.testing_classes import PublishTest
 from tests.lib.assert_classes import DBAssert
+from tests.integration.hosts.nuke.lib import NukeTestClass
 
 log = logging.getLogger("test_publish_in_nuke")
 
 
-class TestPublishInNuke(PublishTest):
+class TestPublishInNuke(NukeTestClass):
     """Basic test case for publishing in Nuke
 
         Uses generic TestCase to prepare fixtures for test data, testing DBs,
@@ -42,43 +39,6 @@ class TestPublishInNuke(PublishTest):
     APP_VARIANT = ""
     PERSIST = True  # True - keep test_db, test_openpype, outputted test files
     TEST_DATA_FOLDER = None
-
-    @pytest.fixture(scope="module")
-    def last_workfile_path(self, download_test_data, output_folder_url):
-        """Get last_workfile_path from source data.
-
-        """
-        source_file_name = "test_project_test_asset_CompositingInNuke_v001.nk"
-        src_path = os.path.join(download_test_data,
-                                "input",
-                                "workfile",
-                                source_file_name)
-        dest_folder = os.path.join(output_folder_url,
-                                   self.PROJECT,
-                                   self.ASSET,
-                                   "work",
-                                   self.TASK)
-        if not os.path.exists(dest_folder):
-            os.makedirs(dest_folder)
-
-        dest_path = os.path.join(dest_folder,
-                                 source_file_name)
-
-        shutil.copy(src_path, dest_path)
-
-        yield dest_path
-
-    @pytest.fixture(scope="module")
-    def startup_scripts(self, monkeypatch_session, download_test_data):
-        """Points Nuke to userSetup file from input data"""
-        startup_path = os.path.join(download_test_data,
-                                    "input",
-                                    "startup")
-        original_nuke_path = os.environ.get("NUKE_PATH", "")
-        monkeypatch_session.setenv("NUKE_PATH",
-                                   "{}{}{}".format(startup_path,
-                                                   os.pathsep,
-                                                   original_nuke_path))
 
     def test_db_asserts(self, dbcon, publish_finished):
         """Host and input data dependent expected results in DB."""
